requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
Deprecated==1.2.18
Flask==3.1.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
redis==5.2.1
rich==13.9.4
rq==2.3.1
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
